+++ src/App.jsx
@@ -1,16 +1,20 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { RequireAuth } from "./middleware/signAuth";
+
 import ProtectedRoutes from "./pages/ProtectedRoutes";
-import Landing from "./pages/ProtectedRoutes/landing";
-import ResumeBuilder from "./pages/ProtectedRoutes/resume";
 import DashBoard from "./pages/ProtectedRoutes/DashBoard";
 import Options from "./pages/ProtectedRoutes/builder/options";
 import StartBuilder from "./pages/ProtectedRoutes/builder/start";
 import Builder from "./pages/ProtectedRoutes/builder/screens";
-import Internship from "./pages/ProtectedRoutes/internship";
+
 import Sign from "./pages/sign";
 import Login from "./pages/sign/login/login";
 import SignUp from "./pages/sign/signup/signup";
+
+import DefaultRoutes from "./pages/DefaultRoutes";
+import Landing from "./pages/DefaultRoutes/landing";
+import Internship from "./pages/DefaultRoutes/internship";
+import ResumeBuilder from "./pages/DefaultRoutes/resume";
 
 function App() {
   return (
@@ -20,6 +24,15 @@ function App() {
           <Route path="/login" element={<Login />} />
           <Route path="/signup" element={<SignUp />} />
         </Route>
+
+        <Route element={<DefaultRoutes />}>
+          <Route path="/" element={<Landing />} />
+          <Route path="/internship" element={<Internship />} />
+          <Route path="/resume" element={<ResumeBuilder />} />
+          <Route path="/blog" element={<div className="h-screen"></div>} />
+          <Route path="/about" element={<div className="h-screen"></div>} />
+        </Route>
+
         <Route
           element={
             <RequireAuth>
@@ -27,15 +40,10 @@ function App() {
             </RequireAuth>
           }
         >
-          <Route path="/" element={<Landing />} />
-          <Route path="/internship" element={<Internship />} />
-          <Route path="/resume" element={<ResumeBuilder />} />
           <Route path="/resume/options" element={<Options />} />
           <Route path="/resume/start" element={<StartBuilder />} />
           <Route path="/resume/builder" element={<Builder />} />
           <Route path="/dashboard/*" element={<DashBoard />} />
-          <Route path="/blog" element={<div></div>} />
-          <Route path="/about" element={<div></div>} />
         </Route>
       </Routes>
     </BrowserRouter>
