+++ src/components/button/index.jsx
@@ -1,3 +1,5 @@
+import { Link } from "react-router-dom";
+
 // Create a custom components for the major buttons in the app
 export function CustomButton({
   primary,
@@ -25,8 +27,11 @@ export function CustomButton({
 
 export function ResumeButton() {
   return (
-    <button className="self-start px-6 py-2 bg-white flex gap-2 items-center justify-center text-primary-500 border-primary-500 border rounded-sm">
+    <Link
+      to="/resume/options"
+      className="self-start px-6 py-2 bg-white flex gap-2 items-center justify-center text-primary-500 border-primary-500 border rounded-sm"
+    >
       Build resume now
-    </button>
+    </Link>
   );
 }
