+++ src/layouts/header/index.jsx
@@ -51,7 +51,7 @@ function Header() {
           : "nav-bar-hidden"
       }
     >
-      <nav className="w-11/12 mx-auto max-w-screen-2xl flex justify-between items-center gap-16 py-3">
+      <nav className="w-11/12 mx-auto max-w-screen-2xl flex justify-between items-center gap-16 py-4">
         <Link to="/" className="relative z-30 nav-icon block">
           <img
             className={
@@ -118,7 +118,11 @@ function Header() {
         <nav className="hidden lg:block">
           <Link
             to="/login"
-            className="w-full text-sm px-6 py-2 text-black-500 border-secondary-500 border rounded-sm"
+            className={`w-full text-sm px-6 py-2 text-black-500 border-secondary-500 border rounded-sm ${
+              offset.darken
+                ? "hover:text-neutral-100 hover:bg-white"
+                : "hover:text-white hover:bg-secondary-500"
+            } transistion duration-500`}
           >
             Log In
           </Link>
