+++ src/pages/DefaultRoutes/landing/signPrompt.jsx
@@ -6,7 +6,7 @@ function SignPrompt() {
     <div className="text-sm">
       <div className="border border-gray-400 rounded-lg p-3 xs:p-6 flex flex-col md:flex-row gap-3 xs:gap-6 items-center justify-between">
         <input
-          className="w-full md:flex-1 bg-primaryBg border border-gray-400 px-4 xs:px-6 py-3 rounded-lg"
+          className="w-full md:flex-1 bg-neutral-1000 border border-gray-400 outline-0 px-4 xs:px-6 py-3 rounded-lg"
           type="text"
           placeholder="Enter your email address"
         />
