+++ src/pages/DefaultRoutes/resume/index.css
@@ -1,0 +1,16 @@
+.scroll img {
+  animation: scrollImage 10s infinite linear;
+}
+
+.scroll img:nth-child(2) {
+  animation-delay: 1s;
+}
+
+@keyframes scrollImage {
+  from {
+    transform: translateX(0%);
+  }
+  to {
+    transform: translateX(-100%);
+  }
+}
