+++ src/pages/ProtectedRoutes/index.jsx
@@ -1,15 +1,11 @@
 import { Outlet } from "react-router-dom";
 import ScrollToTop from "../../components/scrollToTop";
-import Footer from "../../layouts/footer";
-import Header from "../../layouts/header";
 
 function ProtectedRoutes() {
   return (
     <>
       <ScrollToTop />
-      <Header />
       <Outlet />
-      <Footer />
     </>
   );
 }
